add gw rules func to pro
